Move the version.rb file.
