+++ conjur-api.gemspec
@@ -1,5 +1,5 @@
 # -*- encoding: utf-8 -*-
-require File.expand_path('../lib/conjur-api/version', __FILE__)
+require File.expand_path('../lib/conjur/api/version', __FILE__)
 
 Gem::Specification.new do |gem|
   gem.authors       = ["Rafa\305\202 Rzepecki","Kevin Gilpin"]
