+++ lib/conjur/api.rb
@@ -32,7 +32,7 @@ require 'conjur/acts_as_asset'
 require 'conjur/authn-api'
 require 'conjur/authz-api'
 require 'conjur/core-api'
-require 'conjur-api/version'
+require 'conjur/api/version'
 
 class RestClient::Resource
   include Conjur::Escape
