+++ lib/conjur/api/version.rb
